Rename inventoryGroup dir
