+++ frontend/awx/resources/inventories/inventoryGroup/GroupPageAdd.cy.tsx
