+++ frontend/awx/resources/inventories/inventoryGroup/GroupPageAdd.tsx
